# Conflicts:
#	base/allocator/dispatcher/tls.cc
#	starboard/nplb/BUILD.gn
#	starboard/nplb/posix_compliance/posix_prctl_test.cc
#	starboard/shared/modular/cobalt_layer_posix_prctl_abi_wrappers.cc
#	starboard/shared/modular/starboard_layer_posix_prctl_abi_wrappers.cc
#	starboard/shared/modular/starboard_layer_posix_prctl_abi_wrappers.h

+++ ash/test/ash_test_helper.cc
@@ -12,6 +12,7 @@
 #include "ash/app_list/test/app_list_test_helper.h"
 #include "ash/assistant/assistant_controller_impl.h"
 #include "ash/assistant/test/test_assistant_service.h"
+#include "ash/constants/ash_features.h"
 #include "ash/constants/ash_switches.h"
 #include "ash/display/display_configuration_controller_test_api.h"
 #include "ash/display/screen_ash.h"
@@ -20,14 +21,19 @@
 #include "ash/keyboard/test_keyboard_ui.h"
 #include "ash/public/cpp/test/test_keyboard_controller_observer.h"
 #include "ash/public/cpp/test/test_new_window_delegate.h"
+#include "ash/quick_pair/common/fake_quick_pair_browser_delegate.h"
+#include "ash/quick_pair/keyed_service/fake_quick_pair_mediator_factory.h"
+#include "ash/quick_pair/keyed_service/quick_pair_mediator.h"
 #include "ash/session/test_session_controller_client.h"
 #include "ash/shell.h"
 #include "ash/shell_init_params.h"
 #include "ash/style/dark_light_mode_controller_impl.h"
-#include "ash/system/message_center/session_state_notification_blocker.h"
+#include "ash/system/geolocation/test_geolocation_url_loader_factory.h"
 #include "ash/system/model/system_tray_model.h"
+#include "ash/system/notification_center/session_state_notification_blocker.h"
 #include "ash/system/screen_layout_observer.h"
 #include "ash/test/ash_test_views_delegate.h"
+#include "ash/test/login_info.h"
 #include "ash/test/pixel/ash_pixel_test_helper.h"
 #include "ash/test/toplevel_window.h"
 #include "ash/test_shell_delegate.h"
@@ -41,8 +47,14 @@
 #include "base/system/system_monitor.h"
 #include "chromeos/ash/components/audio/cras_audio_handler.h"
 #include "chromeos/ash/components/dbus/audio/cras_audio_client.h"
+#include "chromeos/ash/components/dbus/dlcservice/dlcservice_client.h"
+#include "chromeos/ash/components/dbus/dlcservice/fake_dlcservice_client.h"
 #include "chromeos/ash/components/dbus/rgbkbd/rgbkbd_client.h"
+#include "chromeos/ash/components/dbus/typecd/typecd_client.h"
+#include "chromeos/ash/components/fwupd/fake_fwupd_download_client.h"
 #include "chromeos/ash/components/login/login_state/login_state.h"
+#include "chromeos/ash/services/bluetooth_config/in_process_instance.h"
+#include "chromeos/ash/services/hotspot_config/public/cpp/cros_hotspot_config_test_helper.h"
 #include "chromeos/dbus/power/power_policy_controller.h"
 #include "chromeos/ui/frame/multitask_menu/multitask_menu_nudge_controller.h"
 #include "device/bluetooth/bluetooth_adapter_factory.h"
@@ -56,6 +68,8 @@
 #include "ui/color/color_provider_manager.h"
 #include "ui/display/display_switches.h"
 #include "ui/display/manager/display_manager.h"
+#include "ui/display/manager/test/fake_display_delegate.h"
+#include "ui/display/manager/util/display_manager_test_util.h"
 #include "ui/display/test/display_manager_test_api.h"
 #include "ui/display/util/display_util.h"
 #include "ui/events/gesture_detection/gesture_configuration.h"
@@ -109,7 +123,9 @@ class AshTestHelper::PowerPolicyControllerInitializer {
 
 AshTestHelper::AshTestHelper(ui::ContextFactory* context_factory)
     : AuraTestHelper(context_factory),
-      system_monitor_(std::make_unique<base::SystemMonitor>()) {
+      system_monitor_(std::make_unique<base::SystemMonitor>()),
+      scoped_fake_federated_service_connection_for_test_(
+          &fake_federated_service_connection_) {
   views::ViewsTestHelperAura::SetFallbackTestViewsDelegateFactory(
       &MakeTestViewsDelegate);
 
@@ -136,16 +152,26 @@ AshTestHelper::AshTestHelper(ui::ContextFactory* context_factory)
   // Clears the saved state so that test doesn't use on the wrong
   // default state.
   shell::ToplevelWindow::ClearSavedStateForTest();
+
+  // SimpleGeolocationProvider has to be initialized before
+  // GeolocationController, which is constructed during Shell::Init().
+  SimpleGeolocationProvider::Initialize(
+      base::MakeRefCounted<TestGeolocationUrlLoaderFactory>());
 }
 
 AshTestHelper::~AshTestHelper() {
-  if (app_list_test_helper_)
+  if (app_list_test_helper_) {
     TearDown();
+  }
 
-  // Ensure the next test starts with a null display::Screen.  This must be done
-  // here instead of in TearDown() since some tests test access to the Screen
-  // after the shell shuts down (which they use TearDown() to trigger).
-  ScreenAsh::DeleteScreenForShutdown();
+  SimpleGeolocationProvider::DestroyForTesting();
+
+  if (destroy_screen_) {
+    // Ensure the next test starts with a null display::Screen.  This must be
+    // done here instead of in TearDown() since some tests test access to the
+    // Screen after the shell shuts down (which they use TearDown() to trigger).
+    ScreenAsh::DeleteScreenForShutdown();
+  }
 
   // This should never have a meaningful effect, since either there is no
   // ViewsTestHelperAura instance or the instance is currently in its
@@ -158,7 +184,8 @@ void AshTestHelper::SetUp() {
 }
 
 void AshTestHelper::TearDown() {
-  saved_desk_test_helper_.reset();
+  fwupd_download_client_.reset();
+  saved_desk_test_helper_->Shutdown();
 
   ambient_ash_test_helper_.reset();
 
@@ -176,6 +203,8 @@ void AshTestHelper::TearDown() {
 
   LoginState::Shutdown();
 
+  TypecdClient::Shutdown();
+
   if (create_global_cras_audio_handler_) {
     CrasAudioHandler::Shutdown();
     CrasAudioClient::Shutdown();
@@ -191,10 +220,12 @@ void AshTestHelper::TearDown() {
 
   // Destroy all owned objects to prevent tests from depending on their state
   // after this returns.
+  cros_hotspot_config_test_helper_.reset();
   test_keyboard_controller_observer_.reset();
   session_controller_client_.reset();
+  dlc_service_client_.reset();
   test_views_delegate_.reset();
-  new_window_delegate_provider_.reset();
+  new_window_delegate_.reset();
   bluez_dbus_manager_initializer_.reset();
   floss_dbus_manager_initializer_.reset();
   system_tray_client_.reset();
@@ -203,6 +234,8 @@ void AshTestHelper::TearDown() {
   prefs_provider_.reset();
   statistics_provider_.reset();
   command_line_.reset();
+  quick_pair_browser_delegate_.reset();
+  saved_desk_test_helper_.reset();
 
   // Purge ColorProviderManager between tests so that we don't accumulate
   // ColorProviderInitializers. crbug.com/1349232.
@@ -222,8 +255,9 @@ void AshTestHelper::TearDown() {
 
 aura::Window* AshTestHelper::GetContext() {
   aura::Window* root_window = Shell::GetRootWindowForNewWindows();
-  if (!root_window)
+  if (!root_window) {
     root_window = Shell::GetPrimaryRootWindow();
+  }
   DCHECK(root_window);
   return root_window;
 }
@@ -253,19 +287,14 @@ aura::client::CaptureClient* AshTestHelper::GetCaptureClient() {
 void AshTestHelper::SetUp(InitParams init_params) {
   create_global_cras_audio_handler_ =
       init_params.create_global_cras_audio_handler;
+  create_quick_pair_mediator_ = init_params.create_quick_pair_mediator;
+  destroy_screen_ = init_params.destroy_screen;
+
   if (create_global_cras_audio_handler_) {
     // Create `CrasAudioHandler` for testing since `g_browser_process` is not
     // created in `AshTestBase` tests.
     CrasAudioClient::InitializeFake();
     CrasAudioHandler::InitializeForTesting();
-  }
-
-  // Build `pixel_test_helper_` only for a pixel diff test.
-  if (init_params.pixel_test_init_params) {
-    // Constructing `pixel_test_helper_` sets the locale. Therefore, building
-    // `pixel_test_helper_` before the code that establishes the Ash UI.
-    pixel_test_helper_ = std::make_unique<AshPixelTestHelper>(
-        std::move(*init_params.pixel_test_init_params));
   }
 
   // This block of objects are conditionally initialized here rather than in the
@@ -288,34 +317,56 @@ void AshTestHelper::SetUp(InitParams init_params) {
     }
   }
 
-  if (!RgbkbdClient::Get())
+  if (!RgbkbdClient::Get()) {
     RgbkbdClient::InitializeFake();
-  if (!chromeos::PowerManagerClient::Get())
+  }
+  if (!chromeos::PowerManagerClient::Get()) {
     chromeos::PowerManagerClient::InitializeFake();
+  }
   if (!chromeos::PowerPolicyController::IsInitialized()) {
     power_policy_controller_initializer_ =
         std::make_unique<PowerPolicyControllerInitializer>();
   }
-  if (!NewWindowDelegate::GetInstance()) {
-    new_window_delegate_provider_ =
-        std::make_unique<TestNewWindowDelegateProvider>(
-            std::make_unique<TestNewWindowDelegate>());
+
+  if (!TypecdClient::Get()) {
+    TypecdClient::InitializeFake();
   }
-  if (!views::ViewsDelegate::GetInstance())
+
+  if (!NewWindowDelegate::GetInstance()) {
+    new_window_delegate_ = std::make_unique<TestNewWindowDelegate>();
+  }
+  if (!views::ViewsDelegate::GetInstance()) {
     test_views_delegate_ = MakeTestViewsDelegate();
+  }
+  if (!DlcserviceClient::Get()) {
+    dlc_service_client_ = std::make_unique<FakeDlcserviceClient>();
+  }
+
+  cros_hotspot_config_test_helper_ =
+      std::make_unique<hotspot_config::CrosHotspotConfigTestHelper>(
+          /*use_fake_implementation=*/true);
 
   LoginState::Initialize();
 
   ambient_ash_test_helper_ = std::make_unique<AmbientAshTestHelper>();
+  quick_pair_browser_delegate_ =
+      std::make_unique<quick_pair::FakeQuickPairBrowserDelegate>();
 
   ShellInitParams shell_init_params;
   shell_init_params.delegate = std::move(init_params.delegate);
-  if (!shell_init_params.delegate)
+  if (!shell_init_params.delegate) {
     shell_init_params.delegate = std::make_unique<TestShellDelegate>();
+  }
   shell_init_params.context_factory = GetContextFactory();
   shell_init_params.local_state = init_params.local_state;
   shell_init_params.keyboard_ui_factory =
       std::make_unique<TestKeyboardUIFactory>();
+  if (create_quick_pair_mediator_) {
+    shell_init_params.quick_pair_mediator_factory =
+        std::make_unique<quick_pair::FakeQuickPairMediatorFactory>();
+  }
+  shell_init_params.native_display_delegate =
+      std::make_unique<display::FakeDisplayDelegate>();
   Shell::CreateInstance(std::move(shell_init_params));
   Shell* shell = Shell::Get();
 
@@ -340,15 +391,13 @@ void AshTestHelper::SetUp(InitParams init_params) {
   shell->assistant_controller()->SetAssistant(assistant_service_.get());
 
   shell->system_tray_model()->SetClient(system_tray_client_.get());
-
-  session_controller_client_ = std::make_unique<TestSessionControllerClient>(
-      shell->session_controller(), prefs_provider_.get());
-  session_controller_client_->InitializeAndSetClient();
-  if (init_params.start_session)
-    session_controller_client_->CreatePredefinedUserSessions(1);
+  prefs_provider_ = std::make_unique<TestPrefServiceProvider>();
 
   // Requires the AppListController the Shell creates.
   app_list_test_helper_ = std::make_unique<AppListTestHelper>();
+
+  // SavedDeskTestHelper depends on account.
+  saved_desk_test_helper_ = std::make_unique<SavedDeskTestHelper>();
 
   Shell::GetPrimaryRootWindow()->Show();
   Shell::GetPrimaryRootWindow()->GetHost()->Show();
@@ -379,9 +428,10 @@ void AshTestHelper::SetUp(InitParams init_params) {
   // Tests expect empty wallpaper.
   shell->wallpaper_controller()->CreateEmptyWallpaperForTesting();
 
-  // Move the mouse cursor to far away so that native events don't interfere
-  // with test expectations.
-  Shell::GetPrimaryRootWindow()->MoveCursorTo(gfx::Point(-1000, -1000));
+  // Native events and mouse movements are disabled by
+  // `ui::DisableNativeUiEventDispatchDisabled()`. Just make sure that the the
+  // mouse cursour is not on the screen by default.
+  aura::Env::GetInstance()->SetLastMouseLocation(gfx::Point(-1000, -1000));
   shell->cursor_manager()->EnableMouseEvents();
 
   // Changing GestureConfiguration shouldn't make tests fail. These values
@@ -397,12 +447,26 @@ void AshTestHelper::SetUp(InitParams init_params) {
   AccelerometerReader::GetInstance()->SetECLidAngleDriverStatusForTesting(
       ECLidAngleDriverStatus::NOT_SUPPORTED);
 
-  // Call `StabilizeUIForPixelTest()` after the user session is activated (if
-  // any) in the test setup.
-  if (pixel_test_helper_)
-    StabilizeUIForPixelTest();
+  if (TabletMode::IsBoardTypeMarkedAsTabletCapable()) {
+    shell->tablet_mode_controller()->OnDeviceListsComplete();
+  }
 
-  saved_desk_test_helper_ = std::make_unique<SavedDeskTestHelper>();
+  fwupd_download_client_ = std::make_unique<FakeFwupdDownloadClient>();
+
+  session_controller_client_ = std::make_unique<TestSessionControllerClient>(
+      shell->session_controller(), prefs_provider_.get(),
+      init_params.create_signin_pref_service);
+  session_controller_client_->set_pref_service_must_exist(
+      !init_params.auto_create_prefs_services);
+  session_controller_client_->InitializeAndSetClient();
+
+  // Sign-in after UI is shown.
+  if (init_params.start_session) {
+    // TODO(crbug.com/383441831): Remove Reset();
+    session_controller_client_->Reset();
+
+    SimulateUserLogin({}, AccountId::FromUserEmail("user0@tray"));
+  }
 }
 
 display::Display AshTestHelper::GetSecondaryDisplay() const {
@@ -410,17 +474,20 @@ display::Display AshTestHelper::GetSecondaryDisplay() const {
       .GetSecondaryDisplay();
 }
 
-void AshTestHelper::SimulateUserLogin(const AccountId& account_id,
-                                      user_manager::UserType user_type) {
-  session_controller_client_->AddUserSession(
-      account_id, account_id.GetUserEmail(), user_type);
+AccountId AshTestHelper::SimulateUserLogin(
+    LoginInfo login_info,
+    std::optional<AccountId> opt_account_id,
+    std::unique_ptr<PrefService> pref_service) {
+  AccountId account_id = session_controller_client_->AddUserSession(
+      login_info, opt_account_id, std::move(pref_service));
   session_controller_client_->SwitchActiveUser(account_id);
-  session_controller_client_->SetSessionState(
-      session_manager::SessionState::ACTIVE);
-}
 
-void AshTestHelper::StabilizeUIForPixelTest() {
-  pixel_test_helper_->StabilizeUi();
+  if (login_info.activate_session) {
+    session_controller_client_->SetSessionState(
+        session_manager::SessionState::ACTIVE);
+  }
+
+  return account_id;
 }
 
 }  // namespace ash
